backend/requirements.txt
Django>=5.0.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
mysqlclient>=2.2.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.1
django-filter>=23.5
Pillow>=10.2.0
setuptools

